[backend] update: update product controller, product model, [frontend-admin] feat: basic admin's functions

+++ backend/app/controllers/product.controller.js
@@ -9,7 +9,7 @@ class ProductController {
     async index(req, res, next) {
         try {
             const productModel = new ProductModel();
-            const products = await productModel.getAll();
+            const products = await productModel.getAll(req.query.limit, req.query.offset);
             res.status(StatusCodes.OK).json({
                 status: 'success',
                 data: products,
@@ -32,17 +32,84 @@ class ProductController {
             next(new BadRequestError(error.message));
         }
     }
+    async getCount(req, res, next) {
+        try {
+            const productModel = new ProductModel();
+            const count = await productModel.getCount();
+            res.status(StatusCodes.OK).json({
+                status: 'success',
+                data: count,
+            });
+        }
+        catch(error) {
+            next(new BadRequestError(error.message));
+        }
+    }
+    async getCountByCategory(req, res, next) {
+        try {
+            const productModel = new ProductModel();
+            const count = await productModel.getCountByCategory(req.params.categoryId);
+            res.status(StatusCodes.OK).json({
+                status: 'success',
+                data: count,
+            });
+        }
+        catch(error) {
+            next(new BadRequestError(error.message));
+        }
+    }
+    async getCountByDiscount(req, res, next) {
+        try {
+            const productModel = new ProductModel();
+            const count = await productModel.getCountByDiscount(req.params.discountId);
+            res.status(StatusCodes.OK).json({
+                status: 'success',
+                data: count,
+            });
+        }
+        catch(error) {
+            next(new BadRequestError(error.message));
+        }
+    }
+    async getCountByFilter(req, res, next) {
+        try {
+            const productModel = new ProductModel();
+            const count = await productModel.getCountByFilter(req.query.status, req.query.category);
+            res.status(StatusCodes.OK).json({
+                status: 'success',
+                data: count,
+            });
+        }
+        catch(error) {
+            next(new BadRequestError(error.message));
+        }
+    }
+    async getProductsByFilter(req, res, next) {
+        try {
+            const productModel = new ProductModel();
+            const products = await productModel.getProductsByFilter(req.query.status, req.query.category, req.query.limit, req.query.offset);
+            res.status(StatusCodes.OK).json({
+                status: 'success',
+                data: products,
+            });
+        }
+        catch(error) {
+            next(new BadRequestError(error.message));
+        }
+    }
     async create(req, res, next) {
         uploadProductImages(req, res, async err => {
             if(err instanceof multer.MulterError) {
-                req.files.forEach(async file => {
-                    try {
-                        await unlink(file.path);
-                    }
-                    catch(err) {
-                        return next(new BadRequestError(err.message));
-                    }
-                });
+                if(req.files?.length > 0) {
+                    req.files.forEach(async file => {
+                        try {
+                            await unlink(file.path);
+                        }
+                        catch(err) {
+                            return next(new BadRequestError(err.message));
+                        }
+                    });
+                }
                 return next(new BadRequestError(err.message));
             }
             else if(err) {
@@ -56,7 +123,7 @@ class ProductController {
                 });
                 return next(new BadRequestError(err.message));
             }
-            if(req.files.length > 0) {
+            if(req.files?.length > 0) {
                 req.body['product_images'] = [ ...req.files ];
             }
             try {
@@ -68,14 +135,16 @@ class ProductController {
                 });
             }
             catch(error) {
-                req.files.forEach(async file => {
-                    try {
-                        await unlink(file.path);
-                    }
-                    catch(err) {
-                        return next(new BadRequestError(err.message));
-                    }
-                });
+                if(req.files?.length > 0) {
+                    req.files.forEach(async file => {
+                        try {
+                            await unlink(file.path);
+                        }
+                        catch(err) {
+                            return next(new BadRequestError(err.message));
+                        }
+                    });
+                }
                 next(new BadRequestError(error.message));
             }
         });
@@ -83,14 +152,16 @@ class ProductController {
     async update(req, res, next) {
         uploadProductImages(req, res, async err => {
             if(err instanceof multer.MulterError) {
-                req.files.forEach(async file => {
-                    try {
-                        await unlink(file.path);
-                    }
-                    catch(err) {
-                        return next(new BadRequestError(err.message));
-                    }
-                });
+                if(req.files?.length > 0) {
+                    req.files.forEach(async file => {
+                        try {
+                            await unlink(file.path);
+                        }
+                        catch(err) {
+                            return next(new BadRequestError(err.message));
+                        }
+                    });
+                }
                 return next(new BadRequestError(err.message));
             }
             else if(err) {
@@ -104,7 +175,7 @@ class ProductController {
                 });
                 return next(new BadRequestError(err.message));
             }
-            if(req.files.length > 0) {
+            if(req.files?.length > 0) {
                 req.body['product_images'] = [ ...req.files ];
             }
             try {
@@ -116,14 +187,16 @@ class ProductController {
                 });
             }
             catch(error) {
-                req.files.forEach(async file => {
-                    try {
-                        await unlink(file.path);
-                    }
-                    catch(err) {
-                        return next(new BadRequestError(err.message));
-                    }
-                });
+                if(req.files?.length > 0) {
+                    req.files.forEach(async file => {
+                        try {
+                            await unlink(file.path);
+                        }
+                        catch(err) {
+                            return next(new BadRequestError(err.message));
+                        }
+                    });
+                }
                 next(new BadRequestError(error.message));
             }
         });
